Move filings table to separate page.
