+++ static/js/modules/tables.js
@@ -112,29 +112,6 @@ var filingsColumns = [
   currencyColumn({data: 'total_independent_expenditures', className: 'min-tablet'}),
 ];
 
-var filingsTableColumns = [
-  {
-    data: 'pdf_url',
-    className: 'all',
-    orderable: false,
-    render: function(data, type, row, meta) {
-      var anchor = document.createElement('a');
-      anchor.textContent = 'View filing';
-      anchor.setAttribute('href', data);
-      anchor.setAttribute('target', '_blank');
-      return anchor.outerHTML;
-    }
-  },
-  {data: 'committee_name', className: 'min-desktop', orderable: false},
-  {data: 'candidate_name', className: 'min-desktop', orderable: false},
-  {data: 'amendment_indicator', className: 'min-desktop'},
-  {data: 'report_type_full', className: 'min-desktop', orderable: false},
-  dateColumn({data: 'receipt_date', className: 'min-tablet'}),
-  currencyColumn({data: 'total_receipts', className: 'min-tablet'}),
-  currencyColumn({data: 'total_disbursements', className: 'min-tablet'}),
-  currencyColumn({data: 'total_independent_expenditures', className: 'min-tablet'})
-];
-
 var disbursementPurposeColumns = [
   {data: 'purpose', className: 'all', orderable: false},
   currencyColumn({data: 'total', className: 'all', orderable: false})
@@ -353,12 +330,6 @@ module.exports = {
       var year = $table.attr('data-year');
       var path, query;
       switch ($table.attr('data-type')) {
-        case 'filing-table':
-          initTable($table, $form, 'filings', {}, filingsTableColumns, offsetCallbacks, {
-            // Order by receipt date descending
-            order: [[5, 'desc']],
-          });
-          break;
         case 'filing':
           initTable($table, $form, 'committee/' + committeeId + '/filings', {}, filingsColumns, offsetCallbacks, {
             // Order by receipt date descending
